feat: mocks folder
